Added methods to the Javascript interface for specifying branching priority.
For example, add "branchingPriority([[a,b], [c,d], [e]]);" to the Javascript will
construct a ClaferOption with 1) Clafers a and b as the highest priority, 2) Clafers
c and d as the second highest priority, 3) Clafers e as the third highest priority,
4) the remaining Clafers as the lowest priority.

As usual, use the Javascript class to parse the input file. The parsed object (JavascriptFile)
now contains an additional field for ClaferOption. This ClaferOption object must be passed to
the compiler for the branching priority to take effect.

For example:

   JavascriptFile p = Javascript.readModel(testFile);
   ClaferSolver s = ClaferCompiler.compile(p.getModel(), p.getScope(), p.getOption());