Fix multiplication when bounded by a explicit mulRange.
For example, for the expression a * b with a mulRange of {-2,...,2} will restrict the possibilities of a and b such that the product must lie in this range even if the multiplication is under a conditional. This will omit some potentially valid solutions. With the fix, you may notice more instances than previously if your model contains multiplication.