Implement partial integer analysis for conditional assignments for group cardinalities.
Condition assignments are when a variable is assigned under all alternatives. For example, consider the following model.

A -> int
xor B
    C
        [ A = 3 ]
    D
        [ A = 4 ]

In this model, A must be assigned to either 3 or 4. Note that if one of the constraints above were removed, we would know nothing about the values of A.